Extract backend provisioning helpers
